refactor: use ACF for commands and split into multiple classes
