+++ build.gradle
@@ -1,4 +1,5 @@
 plugins {
+    id 'com.gradleup.shadow' version '8.3.0'
     id 'java'
     id 'com.palantir.git-version' version '3.0.0'
 }
@@ -17,10 +18,13 @@ repositories {
     maven { url 'https://repo.papermc.io/repository/maven-public/' }
     maven { url 'https://oss.sonatype.org/content/groups/public/' }
     maven { url 'https://repo.extendedclip.com/content/repositories/placeholderapi/' }
+    maven { url 'https://repo.aikar.co/content/groups/aikar/' }
 }
 
 dependencies {
     compileOnly("io.papermc.paper:paper-api:${paperapi}")
+
+    implementation("co.aikar:acf-paper:${paperacf}")
 
     compileOnly("net.kyori:adventure-text-minimessage:${adventure}")
 
@@ -30,7 +34,23 @@ dependencies {
     annotationProcessor("org.projectlombok:lombok:${lombok}")
 }
 
-def targetJavaVersion = 21
+shadowJar {
+    relocate 'co.aikar.commands', 'org.notionsmp.dreiMotd.libs.acf'
+    relocate 'co.aikar.locales', 'org.notionsmp.dreiMotd.libs.locales'
+
+    exclude('org.projectlombok:lombok:.*', 'net.kyori:.*')
+    exclude 'META-INF/**'
+
+    minimize()
+
+    archiveClassifier.set('')
+}
+
+jar.enabled = false
+
+tasks.build.dependsOn tasks.shadowJar
+
+def targetJavaVersion = Integer.parseInt(project.targetJavaVersion)
 java {
     def javaVersion = JavaVersion.toVersion(targetJavaVersion)
     sourceCompatibility = javaVersion
