+++ gradle.properties
@@ -8,5 +8,6 @@ targetJavaVersion=21
 paperapi=1.21-R0.1-SNAPSHOT
 adventure=4.19.0
 lombok=1.18.36
+paperacf=0.5.1-SNAPSHOT
 
 placeholderapi=2.11.5
